requirements/base.txt
Django==2.0.7
django-environ==0.4.5
djangorestframework==3.8.2
psycopg2==2.7.5
django-registration==2.4.1
